refactor: styled-component를 scss로 마이그레이션하라
- remove styled-component dependencies
- install clsx dependency
- 모든 컴포넌트에 scss 적용 및 styled-component 삭제
