--- @types/styled-components.d.ts
@@ -1,16 +1,0 @@
-import type { CSSProp } from 'styled-components';
-
-import lightTheme from '../styles/theme';
-
-type ThemeType = typeof lightTheme;
-
-declare module 'styled-components' {
-  // eslint-disable-next-line @typescript-eslint/no-empty-interface
-  export interface DefaultTheme extends ThemeType {}
-}
-
-declare module 'react' {
-  interface DOMAttributes<T> {
-    css?: CSSProp<T>;
-  }
-}
