+++ components/about/searchCountryHeader/index.module.scss
@@ -1,5 +1,3 @@
-@import "/styles/main.scss";
-
 .buttonWrapper {
   width: calc(100% - 40px);
   position: fixed;
