+++ components/map/PlaceBottomSheetQ.tsx
@@ -2,8 +2,6 @@ import { useEffect, useState } from 'react';
 import Sheet from 'react-modal-sheet';
 
 import { PlaceGeometry } from 'lib/types/google.maps';
-import styled from 'styled-components';
-import { titleLargeFont } from 'styles/fontStyles';
 
 type Props = {
   placeResult: PlaceGeometry[];
@@ -12,6 +10,7 @@ type Props = {
 const snapPoints = [-50, 0.5, 200, 0];
 const initialSnap = 1;
 
+// TODO - 추후 삭제
 function PlaceBottomSheetQ({ placeResult }: Props) {
   const [open, setOpen] = useState<boolean>(false);
 
@@ -34,17 +33,17 @@ function PlaceBottomSheetQ({ placeResult }: Props) {
       <Sheet.Container>
         <Sheet.Header />
         <Sheet.Content>
-          <PlaceList>
+          <ul>
             {placeResult.map(({
               name, place_id, rating, user_ratings_total,
             }) => (
-              <PlaceItem key={place_id}>
+              <li key={place_id}>
                 <div className="place-name">{name}</div>
                 <div>{rating}</div>
                 <div>{user_ratings_total}</div>
-              </PlaceItem>
+              </li>
             ))}
-          </PlaceList>
+          </ul>
         </Sheet.Content>
       </Sheet.Container>
       <Sheet.Backdrop />
@@ -54,25 +53,3 @@ function PlaceBottomSheetQ({ placeResult }: Props) {
 }
 
 export default PlaceBottomSheetQ;
-
-const PlaceList = styled.ul`
-  list-style: none;
-  margin: 0px;
-  padding: 0px;
-`;
-
-const PlaceItem = styled.li`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  padding: 16px 24px;
-  gap: 4px;
-  padding: 16px 24px;
-  border-bottom: 1px solid ${({ theme }) => theme.gray200};
-
-  & > div.place-name {
-    ${titleLargeFont({ fontWeight: 600 })};
-    letter-spacing: -0.01em;
-    color: ${({ theme }) => theme.black};
-  }
-`;
