+++ components/map/placeBottomSheet/index.module.scss
@@ -1,0 +1,5 @@
+.placeList {
+  list-style: none;
+  margin: 0px;
+  padding: 0px;
+}
