+++ components/map/placeResultMarker/index.tsx
