+++ components/map/searchInput/index.module.scss
@@ -1,0 +1,58 @@
+@import "/styles/main.scss";
+
+.inputWrapper {
+  position: relative;
+  display: flex;
+  flex-direction: row;
+  top: 24px;
+  right: 24px;
+  position: absolute;
+  width: calc(100% - 48px);
+
+  &:focus-within {
+    & div > svg.searchIcon > path {
+      fill: color("purple.500");
+    }
+  }
+}
+
+.searchIconWrapper {
+  cursor: pointer;
+  position: absolute;
+  display: flex;
+  top: 16px;
+  left: 20px;
+}
+
+.searchInput {
+  @include text("headline/500", "black");
+  letter-spacing: -0.012em;
+
+  background: color("white");
+  box-shadow: 0px 4px 0px rgba(19, 17, 24, 0.16);
+  outline: none;
+  width: 100%;
+  height: 56px;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  padding: 12px 60px;
+  border: 2px solid color("black");
+  border-radius: 56px;
+
+  &::placeholder {
+    color: color("gray.400");
+  }
+
+  &:focus {
+    caret-color: color("purple.500");
+  }
+}
+
+.menuIconWrapper {
+  display: flex;
+  position: absolute;
+  top: 16px;
+  right: 20px;
+  cursor: pointer;
+}
