+++ package.json
@@ -30,15 +30,14 @@
   "homepage": "https://github.com/mbti-nf-team/unnamed-k-place-frontend#readme",
   "dependencies": {
     "@react-google-maps/api": "^2.18.1",
+    "clsx": "^1.2.1",
     "framer-motion": "^10.2.2",
     "next": "^13.2.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-modal-sheet": "^1.10.1",
     "react-spring-bottom-sheet": "3.5.0-alpha.0",
-    "sass": "^1.59.3",
-    "styled-components": "beta",
-    "styled-normalize": "^8.0.7"
+    "sass": "^1.59.3"
   },
   "devDependencies": {
     "@commitlint/cli": "^17.4.3",
