--- styles/theme.ts
@@ -1,53 +1,0 @@
-const lightTheme = {
-  /* white */
-  white: '#ffffff',
-
-  /* black */
-  black: '#131118ff',
-  blackAlpha096: '#131118f5',
-  blackAlpha088: '#131118e0',
-  blackAlpha080: '#131118cc',
-  blackAlpha072: '#131118b8',
-  blackAlpha064: '#131118a3',
-  blackAlpha056: '#1311188f',
-  blackAlpha048: '#1311187a',
-  blackAlpha040: '#13111866',
-  blackAlpha032: '#13111852',
-  blackAlpha024: '#1311183d',
-  blackAlpha020: '#13111833',
-  blackAlpha018: '#1311182e',
-  blackAlpha016: '#13111829',
-  blackAlpha014: '#13111824',
-  blackAlpha012: '#1311181f',
-  blackAlpha010: '#1311181a',
-  blackAlpha008: '#13111814',
-  blackAlpha006: '#1311180f',
-  blackAlpha004: '#1311180a',
-  blackAlpha002: '#13111805',
-
-  /* gray */
-  gray000: '#fdfcfd',
-  gray100: '#f4f2f4',
-  gray200: '#e9e8ea',
-  gray300: '#dcdbdd',
-  gray400: '#c8c7cb',
-  gray500: '#908e96',
-  gray600: '#6f6e77',
-  gray700: '#504f57',
-  gray800: '#34343a',
-  gray900: '#1a1523',
-
-  /* purple */
-  purple000: '#fdfcfe',
-  purple100: '#f5f2ff',
-  purple200: '#E1DDFE',
-  purple300: '#C4BAFD',
-  purple400: '#9487F8',
-  purple500: '#6B56F0',
-  purple600: '#523ACB',
-  purple700: '#41319B',
-  purple800: '#382E76',
-  purple900: '#211943',
-} as const;
-
-export default lightTheme;
